# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem Mahad (Absensi, Hafalan, dll)",
        "title": "Mahad Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/login/mahasantri": {
            "post": {
                "description": "Melakukan login untuk mahasantri dengan NIM dan password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login Mahasantri",
                "parameters": [
                    {
                        "description": "Data login Mahasantri",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginMahasantriRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessLoginMahasantriExample"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorLoginMahasantriExample"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorLoginMahasantriExample"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login/mentor": {
            "post": {
                "description": "Melakukan login untuk mentor dengan email dan password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login Mentor",
                "parameters": [
                    {
                        "description": "Data login Mentor",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginMentorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessLoginMentorExample"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorLoginMentorExample"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorLoginMentorExample"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data user yang sedang login (baik Mentor atau Mahasantri)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessGetCurrentUserExample"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorGetCurrentUserExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorGetCurrentUserExample"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register/mahasantri": {
            "post": {
                "description": "Mendaftarkan akun Mahasantri baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register Mahasantri",
                "parameters": [
                    {
                        "description": "Data pendaftaran Mahasantri",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterMahasantriRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessExample"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorExample"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorExample"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register/mentor": {
            "post": {
                "description": "Mendaftarkan akun Mentor baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register Mentor",
                "parameters": [
                    {
                        "description": "Data pendaftaran Mentor",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterMentorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.SuccessMentorExample"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorMentorExample"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorMentorExample"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginMahasantriRequest": {
            "type": "object",
            "required": [
                "nim",
                "password"
            ],
            "properties": {
                "nim": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "dto.LoginMentorRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterMahasantriRequest": {
            "type": "object",
            "required": [
                "gender",
                "jurusan",
                "mentor_id",
                "nama",
                "nim",
                "password"
            ],
            "properties": {
                "gender": {
                    "type": "string",
                    "enum": [
                        "L",
                        "P"
                    ]
                },
                "jurusan": {
                    "type": "string"
                },
                "mentor_id": {
                    "description": "Mentor yang membimbing",
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                },
                "nim": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "dto.RegisterMentorRequest": {
            "type": "object",
            "required": [
                "email",
                "gender",
                "nama",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "L",
                        "P"
                    ]
                },
                "nama": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "utils.ErrorExample": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.ErrorGetCurrentUserExample": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "Unauthorized"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.ErrorLoginMahasantriExample": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "Invalid NIM or password"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.ErrorLoginMentorExample": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "Invalid email or password"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "utils.ErrorMentorExample": {
            "type": "object",
            "properties": {
                "error": {},
                "message": {
                    "type": "string",
                    "example": "Invalid request body"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "utils.SuccessExample": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Mahasantri registered successfully"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessGetCurrentUserExample": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "User data retrieved"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessLoginMahasantriExample": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessLoginMentorExample": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "utils.SuccessMentorExample": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Mentor registered successfully"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}